// src/app/api/generated/model/create-espacio-request.ts
/**
 * AKINE API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Datos para dar de alta un espacio en la sede
 */
export interface CreateEspacioRequest { 
    /**
     * Personas simultaneas que admite. Si se omite, 1, que es el caso de un box individual
     */
    capacidad?: number;
    /**
     * Nombre operativo del espacio. Unico entre los espacios VIGENTES de la sede: el nombre de uno dado de baja se puede reusar
     */
    name: string;
    /**
     * Observacion operativa libre. Nunca contenido clinico
     */
    notes?: string;
    /**
     * Clasificacion fisica del recurso. Si se omite, BOX. NO habilita servicios ni define roles: eso lo resuelve la habilitacion por Oferta de Servicio (RN-M04-007)
     */
    tipo?: CreateEspacioRequestTipoEnum;
    /**
     * Instante UTC desde el que el espacio esta en servicio. Si se omite, el momento del alta. Es la ventana OPERATIVA, distinta de la baja logica
     */
    validFrom?: string;
    /**
     * Instante UTC hasta el que esta en servicio, EXCLUSIVO. Omitirlo significa sin fin previsto, que es lo normal
     */
    validUntil?: string;
}
export enum CreateEspacioRequestTipoEnum {
    BOX = 'BOX',
    GIMNASIO = 'GIMNASIO',
    GABINETE = 'GABINETE',
    SALA_GRUPAL = 'SALA_GRUPAL',
    PILETA = 'PILETA',
    OTRO = 'OTRO'
};